Make sure JSContext is only ever called from the main thread
I don't know how this hasn't caused much more issue for us.
SafariExtensionHandler.messageReceived() always runs on the main thread,
but if you call e.g. SFSafariApplication.getActiveWindow, the callback
runs on any thread, most likely a background one. JSContext access,
meanwhile, is only ever supposed to happen from the thread that created
it.

I've been getting strange hanging of the extension when running
development builds, and when pausing in the debugger, seeing that we are
in some async await section that never returns. This change fixes that.